Added Jekyll + Hyde theme
Woot!

To use:

Install Jekyll

```
gem install jekyll
```

cd to your working directory and checkout the `gh-pages` branch

Don't modify anything in the `_site` directory (it's auto-generated by
Jekyll so changes will be overwritten)

Launch jekyll server in "watch" mode:

```
jekyll server --watch --baseurl ""
```

The `--baseurl` option is necessary to overwrite the option in
`config.yml`, which is necessary for GitHub to do the right thing.